moved test to a more appropiate place
